Rename com.paypal.android.ui.paypal package to com.paypal.android.ui.paypalnative and fix complete order typo.
