+++ Demo/src/main/java/com/paypal/android/ui/paypalnative/PayPalNativeViewModel.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.paypal
+package com.paypal.android.ui.paypalnative
 
 import android.app.Application
 import android.widget.Toast
